refactor(daemon): extract prompt entity context
